+++ routes/sponsoring.php
@@ -5,7 +5,6 @@ use App\Http\Controllers\Sponsoring\BackerOverview;
 use App\Http\Controllers\Sponsoring\ContractDetail;
 use App\Http\Controllers\Sponsoring\Overview;
 use App\Http\Controllers\Sponsoring\PackageOverview;
-use App\Http\Controllers\Sponsoring\PeriodAdOptionOverview;
 use App\Livewire\Sponsoring\AdOptionCreate;
 use App\Livewire\Sponsoring\AdOptionEdit;
 use App\Livewire\Sponsoring\BackerCreate;
@@ -13,6 +12,7 @@ use App\Livewire\Sponsoring\BackerEdit;
 use App\Livewire\Sponsoring\ContractEdit;
 use App\Livewire\Sponsoring\PackageCreate;
 use App\Livewire\Sponsoring\PackageEdit;
+use App\Livewire\Sponsoring\PeriodAdOptionOverview;
 use App\Livewire\Sponsoring\PeriodBackerOverview;
 use App\Livewire\Sponsoring\PeriodCreate;
 use App\Livewire\Sponsoring\PeriodEdit;
@@ -30,7 +30,7 @@ Route::middleware(['auth', 'permission:'.Contract::SPONSORING_SHOW_PERMISSION.'|
         ->name('sponsoring.package.index');
     Route::get('/sponsoring/period/backer/{period}', PeriodBackerOverview::class)
         ->name('sponsoring.period.backer.overview');
-    Route::get('/sponsoring/period/adOption/{period}', [PeriodAdOptionOverview::class, 'index'])
+    Route::get('/sponsoring/period/adOption/{period}', PeriodAdOptionOverview::class)
         ->name('sponsoring.period.adOption.overview');
     Route::get('/sponsoring/contract/{contract}', [ContractDetail::class, 'index'])
         ->name('sponsoring.contract.detail');